code modularizing
