+++ package.json
@@ -15,6 +15,7 @@
     "@mui/material": "^6.1.10",
     "@mui/x-data-grid": "^7.23.1",
     "framer-motion": "^11.13.1",
+    "lodash": "^4.17.21",
     "next": "15.0.3",
     "next-auth": "^4.24.10",
     "react": "^18.2.0",
