+++ src/app/page.tsx
@@ -1,11 +1,7 @@
-import PageName from "../components/pagename";
-
-export default function Home() {
-
-  return (
-    <div>
-      <PageName PageName="Home" />
+export default function PageName() {
+  return(
+    <div className="m-16 text-lg border-b w-[70rem] border-[#CBD0DD]" >
+        <h1 className="pb-2" > Home </h1>
     </div>
-  );
-
+  )
 }
