+++ src/components/clients/AddClientDialog.tsx
@@ -13,7 +13,7 @@ import {
 } from "@mui/material";
 import { useState } from "react";
 import ArrowBackIcon from "@mui/icons-material/ArrowBack";
-import { Client } from "src/types/Clients/client";
+import { Client } from "src/types/client";
 
 // Import Client interface
