+++ src/types/searchParams.ts
@@ -1,0 +1,3 @@
+export interface SearchParams {
+    name?: string
+}
